# openapi.yaml
openapi: 3.1.0
info:
  title: VoiceBlender API
  description: >
    VoiceBlender bridges SIP and WebRTC voice calls with multi-party audio
    mixing, real-time speech-to-text, text-to-speech, AI agent integration,
    recording, and webhook-based event delivery.
  version: 1.0.0
  license:
    name: MIT
servers:
  - url: http://localhost:8080/v1
    description: Local development server
...
paths:
...
  /legs/{id}/stt:
    parameters:
      - $ref: '#/components/parameters/LegId'
    post:
      operationId: sttLeg
      summary: Start real-time speech-to-text on a leg
      tags: [Legs]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/STTRequest'
      responses:
        '200':
          description: STT started
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: stt_started
                  leg_id:
                    type: string
                required: [status, leg_id]
        '404':
          $ref: '#/components/responses/LegNotFound'
        '409':
          description: Leg not connected, STT already running, or no audio reader
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: No ElevenLabs API key provided
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      operationId: stopSTTLeg
      summary: Stop speech-to-text on a leg
      tags: [Legs]
      responses:
        '200':
          description: STT stopped
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example:
                status: stt_stopped
        '404':
          description: No STT in progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /rooms/{id}/tts:
    parameters:
      - $ref: '#/components/parameters/RoomId'
    post:
      operationId: ttsRoom
      summary: Synthesize speech and play it into a room
      tags: [Rooms]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TTSRequest'
      responses:
        '200':
          description: TTS playback started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaybackResponse'
        '400':
          description: Invalid JSON, missing text/voice, or volume out of range
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/RoomNotFound'
        '409':
          description: Room has no participants
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: No ElevenLabs API key provided
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /rooms/{id}/stt:
    parameters:
      - $ref: '#/components/parameters/RoomId'
    post:
      operationId: sttRoom
      summary: Start speech-to-text on all room participants
      tags: [Rooms]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/STTRequest'
      responses:
        '200':
          description: STT started
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: stt_started
                  room_id:
                    type: string
                  leg_ids:
                    type: array
                    items:
                      type: string
                required: [status, room_id, leg_ids]
        '404':
          $ref: '#/components/responses/RoomNotFound'
        '409':
          description: STT already running or room has no participants
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: No ElevenLabs API key provided
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      operationId: stopSTTRoom
      summary: Stop speech-to-text on a room
      tags: [Rooms]
      responses:
        '200':
          description: STT stopped
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example:
                status: stt_stopped
        '404':
          description: No STT in progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /webhooks:
    post:
      operationId: registerWebhook
      summary: Register a webhook
      tags: [Webhooks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookRequest'
      responses:
        '201':
          description: Webhook registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'
        '400':
          description: Invalid JSON or missing URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      operationId: listWebhooks
      summary: List all registered webhooks
      tags: [Webhooks]
      responses:
        '200':
          description: Array of webhooks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Webhook'
  /webhooks/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: Webhook ID
    delete:
      operationId: deleteWebhook
      summary: Unregister a webhook
      tags: [Webhooks]
      responses:
        '200':
          description: Webhook deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example:
                status: deleted
        '404':
          description: Webhook not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

# ════════════════════════════════════════════════════════════════════════════
components:
  parameters:
    LegId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Leg ID
    RoomId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Room ID
...
  responses:
    LegNotFound:
      description: Leg not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RoomNotFound:
      description: Room not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Webhook:
      type: object
      properties:
        id:
          type: string
          description: Webhook identifier (UUID)
        url:
          type: string
          format: uri
          description: Delivery URL
        secret:
          type: string
          description: HMAC-SHA256 signing secret (omitted if not set)
      required: [id, url]
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
      required: [error]
    StatusResponse:
      type: object
      properties:
        status:
          type: string
      required: [status]

    # ── Request bodies ──────────────────────────────────────────────────
...
    TTSRequest:
      type: object
      properties:
        text:
          type: string
          description: Text to synthesize
        voice:
          type: string
          description: ElevenLabs voice name or ID
        model_id:
          type: string
          description: ElevenLabs model ID
        volume:
          type: integer
          description: Volume adjustment in dB (-8 to 8)
          minimum: -8
          maximum: 8
          default: 0
        api_key:
          type: string
          description: ElevenLabs API key override
      required: [text, voice]
    STTRequest:
      type: object
      properties:
        language:
          type: string
          description: Language code (e.g. "en", "es")
        partial:
          type: boolean
          description: Emit partial (non-final) transcripts
          default: false
        api_key:
          type: string
          description: ElevenLabs API key override
...
    WebhookRequest:
      type: object
      properties:
        url:
          type: string
          format: uri
          description: Webhook delivery URL
        secret:
          type: string
          description: HMAC-SHA256 signing secret
      required: [url]
...
    PlaybackResponse:
      type: object
      properties:
        playback_id:
          type: string
          description: Unique playback identifier
        status:
          type: string
          example: playing
      required: [playback_id, status]
...